Update patch set 10
Patch Set 10: Code-Review-1

(1 comment)

I think this is the first case in Acumos where an AT&T group took third-party open-source software, modified it, and is now offering it for distribution to the world via Gerrit.  We need to seek guidance on the correct way to represent the license, how to give credit to the original author(s) and so on.  I don't have the necessary experience but want to learn.

Patch-set: 10
Label: Code-Review=-1